Remove import cycle
The cycle did not really cause any problems but disabling the warnings
was ugly and so some update code was moved to autoUpdateActions.js.
