+++ src/launcher/actions/appsActions.js
@@ -38,9 +38,6 @@ import { join } from 'path';
 import { ipcRenderer, remote } from 'electron';
 import { ErrorDialogActions } from 'pc-nrfconnect-shared';
 
-// eslint-disable-next-line import/no-cycle
-import { removeSource } from './settingsActions';
-
 const net = remote.require('../main/net');
 const fs = remote.require('fs');
 
@@ -180,19 +177,19 @@ function upgradeOfficialAppErrorAction() {
     };
 }
 
-function downloadLatestAppInfoAction() {
+export function downloadLatestAppInfoAction() {
     return {
         type: DOWNLOAD_LATEST_APP_INFO,
     };
 }
 
-function downloadLatestAppInfoSuccessAction() {
+export function downloadLatestAppInfoSuccessAction() {
     return {
         type: DOWNLOAD_LATEST_APP_INFO_SUCCESS,
     };
 }
 
-function downloadLatestAppInfoErrorAction() {
+export function downloadLatestAppInfoErrorAction() {
     return {
         type: DOWNLOAD_LATEST_APP_INFO_ERROR,
     };
@@ -334,40 +331,6 @@ export function loadOfficialApps(appName, appSource) {
                 dispatch(loadOfficialAppsError());
                 dispatch(ErrorDialogActions.showDialog('Unable to load apps: '
                     + `${error.message}`));
-            });
-    };
-}
-
-export function downloadLatestAppInfo(options = { rejectIfError: false }) {
-    return dispatch => {
-        dispatch(downloadLatestAppInfoAction());
-
-        return mainApps.downloadAppsJsonFiles()
-            .then(() => mainApps.generateUpdatesJsonFiles())
-            .then(() => dispatch(downloadLatestAppInfoSuccessAction()))
-            .then(() => dispatch(loadOfficialApps()))
-            .catch(error => {
-                dispatch(downloadLatestAppInfoErrorAction());
-                if (options.rejectIfError) {
-                    throw error;
-                } else if (net.isResourceNotFound(error)) {
-                    dispatch(ErrorDialogActions.showDialog(
-                        `Unable to retrieve the source “${error.cause.name}” from ${error.cause.url}. \n\n`
-                        + 'This is usually caused by outdated app sources in the settings, '
-                        + 'where the sources files was removed from the server.',
-                        {
-                            'Remove source': () => {
-                                dispatch(removeSource(error.cause.name));
-                                dispatch(ErrorDialogActions.hideDialog());
-                            },
-                            Cancel: () => {
-                                dispatch(ErrorDialogActions.hideDialog());
-                            },
-                        },
-                    ));
-                } else {
-                    dispatch(ErrorDialogActions.showDialog(`Unable to download latest app info: ${error.message}`));
-                }
             });
     };
 }
