+++ src/launcher/actions/autoUpdateActions.js
@@ -38,6 +38,9 @@ import { remote } from 'electron';
 import log from 'electron-log';
 import { ErrorDialogActions } from 'pc-nrfconnect-shared';
 
+import * as AppsActions from './appsActions';
+import * as SettingsActions from './settingsActions';
+
 export const AUTO_UPDATE_CHECK = 'AUTO_UPDATE_CHECK';
 export const AUTO_UPDATE_AVAILABLE = 'AUTO_UPDATE_AVAILABLE';
 export const AUTO_UPDATE_POSTPONE = 'AUTO_UPDATE_POSTPONE';
@@ -47,7 +50,10 @@ export const AUTO_UPDATE_DOWNLOAD_CANCELLED = 'AUTO_UPDATE_DOWNLOAD_CANCELLED';
 export const AUTO_UPDATE_DOWNLOADING = 'AUTO_UPDATE_DOWNLOADING';
 export const AUTO_UPDATE_ERROR = 'AUTO_UPDATE_ERROR';
 
+const mainApps = remote.require('../main/apps');
+const net = remote.require('../main/net');
 const { autoUpdater, CancellationToken } = remote.require('../main/autoUpdate');
+
 const isWindows = process.platform === 'win32';
 const isMac = process.platform === 'darwin';
 
@@ -181,4 +187,51 @@ export function cancelDownload() {
                 + 'No download is in progress.'));
         }
     };
+}
+
+export function downloadLatestAppInfo(options = { rejectIfError: false }) {
+    return dispatch => {
+        dispatch(AppsActions.downloadLatestAppInfoAction());
+
+        return mainApps.downloadAppsJsonFiles()
+            .then(() => mainApps.generateUpdatesJsonFiles())
+            .then(() => dispatch(AppsActions.downloadLatestAppInfoSuccessAction()))
+            .then(() => dispatch(AppsActions.loadOfficialApps()))
+            .catch(error => {
+                dispatch(AppsActions.downloadLatestAppInfoErrorAction());
+                if (options.rejectIfError) {
+                    throw error;
+                } else if (net.isResourceNotFound(error)) {
+                    dispatch(ErrorDialogActions.showDialog(
+                        `Unable to retrieve the source “${error.cause.name}” from ${error.cause.url}. \n\n`
+                        + 'This is usually caused by outdated app sources in the settings, '
+                        + 'where the sources files was removed from the server.',
+                        {
+                            'Remove source': () => {
+                                dispatch(SettingsActions.removeSource(error.cause.name));
+                                dispatch(ErrorDialogActions.hideDialog());
+                            },
+                            Cancel: () => {
+                                dispatch(ErrorDialogActions.hideDialog());
+                            },
+                        },
+                    ));
+                } else {
+                    dispatch(ErrorDialogActions.showDialog(`Unable to download latest app info: ${error.message}`));
+                }
+            });
+    };
+}
+
+export function checkForUpdatesManually() {
+    return dispatch => (
+        dispatch(downloadLatestAppInfo({ rejectIfError: true }))
+            .then(() => {
+                dispatch(checkForCoreUpdates());
+                dispatch(SettingsActions.showUpdateCheckCompleteDialog());
+            })
+            .catch(error => (
+                dispatch(ErrorDialogActions.showDialog(`Unable to check for updates: ${error.message}`))
+            ))
+    );
 }
